+++ src/DemoProject/SharedCode/CustomConsoles/ShapesConsole.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.Xna.Framework;
-using ColorHelper = Microsoft.Xna.Framework.Color;
 
 using Console = SadConsole.Console;
 using SadConsole;
@@ -16,10 +15,13 @@ namespace StarterProject.CustomConsoles
             UseKeyboard = false;
 
             DrawLine(new Point(2, 2), new Point(Width - 4, 2), Color.Yellow, glyph: '=');
-            DrawBox(new Rectangle(2, 4, 6, 6), Color.Yellow, glyph: '=');
-            DrawBox(new Rectangle(9, 4, 6, 6), Color.Yellow, connectedLineStyle: ConnectedLineThin);
-            DrawBox(new Rectangle(16, 4, 6, 6), Color.Yellow, connectedLineStyle: ConnectedLineThick);
-            DrawBox(new Rectangle(23, 4, 6, 6), Color.Black, Color.Yellow, connectedLineStyle: ConnectedLineThick, fill: true);
+            DrawBox(new Rectangle(2, 4, 6, 6), new Cell(Color.Yellow, Color.Black, '='));
+            DrawBox(new Rectangle(9, 4, 6, 6), new Cell(Color.Yellow, Color.Black, '='), connectedLineStyle: ConnectedLineThin);
+            DrawBox(new Rectangle(16, 4, 6, 6), new Cell(Color.Yellow, Color.Black, '='), connectedLineStyle: ConnectedLineThick);
+            DrawBox(new Rectangle(23, 4, 6, 6), new Cell(Color.Black, Color.Yellow, '='), new Cell(Color.Black, Color.Yellow, 0), connectedLineStyle: ConnectedLineThick);
+
+            DrawCircle(new Rectangle(2, 12, 16, 10), new Cell(Color.White, Color.Black, 176));
+            DrawCircle(new Rectangle(19, 12, 16, 10), new Cell(Color.White, Color.Black, 176), new Cell(Color.Green, Color.Black, 178));
 
             IsVisible = false;
         }